add rudimentary login system

+++ tests/api.test.ts
@@ -26,10 +26,43 @@ const userSession = {
   }],
 }
 
+let token: string;
+describe('login', () => {
+    test('POST', async () => {
+      const res = await api.post('/api/login')
+        .send({password: "TEST_PASSWORD"})
+        .expect(200)
+        .expect('Content-type', /application\/json/)
+      expect(res.body.token).toBeDefined()
+      token = res.body.token;
+    })
+    test('POST invalid', async () => {
+      const res = await api.post('/api/login')
+        .send({password: "WRONG"})
+        .expect(401)
+        .expect('Content-type', /application\/json/)
+      expect(res.body.error).toBe('invalid login')
+    })
+    test('unauthenticated', async () => {
+      const res = await api.get('/api/images/ANYWHERE')
+        .expect(401)
+        .expect('Content-type', /application\/json/)
+      expect(res.body.error).toBe('authorization not provided')
+    })
+    test('failed authentication', async () => {
+      const res = await api.get('/api/sessions/ANYWHERE')
+        .auth('INVALID', {type: 'bearer'})
+        .expect(401)
+        .expect('Content-type', /application\/json/)
+      expect(res.body.error).toBe('invalid JWT')
+    })
+})
+
 describe('images', () => {
   const REMOTE_IMGPATH = "/api/images/f4/88/f488b46abf2ab1f02877c0090659de2f.png"
+  const req = () => api.post('/api/images').auth(token, {type: 'bearer'})
   test('POST', async () => {
-    const res = await api.post('/api/images')
+    const res = await req()
       .set('Content-type', 'multipart/form-data')
       .attach('img', 'tests/0.png')
       .expect(200)
@@ -37,7 +70,7 @@ describe('images', () => {
     userSession.imgs[0].src = res.body.url;
   })
   test('POST duplicate', async () => {
-    const res = await api.post('/api/images')
+    const res = await req()
       .set('Content-type', 'multipart/form-data')
       .attach('img', 'tests/0.png')
       .expect(409)
@@ -45,16 +78,18 @@ describe('images', () => {
   })
   test('GET', async () => {
     const res = await api.get(REMOTE_IMGPATH)
+      .auth(token, {type: 'bearer'})
       .expect(200)
       .expect('Content-type', 'image/png')
-    expect(res).toBeDefined()
+    expect(res).toBeDefined();
   })
 })
 
 describe('sessions', () => {
   let sess_id: string;
+  const auth = (req: supertest.Test) => req.auth(token, {type: 'bearer'})
   test('POST /', async () => {
-    const res = await api.post('/api/sessions')
+    const res = await auth(api.post('/api/sessions'))
       .send(userSession)
       .expect(200)
       .expect('Content-type', /application\/json/)
@@ -63,7 +98,7 @@ describe('sessions', () => {
     sess_id = res.body.id;
   })
   test('GET /', async () => {
-    const res = await api.get('/api/sessions')
+    const res = await auth(api.get('/api/sessions'))
       .expect(200)
       .expect('Content-type', /application\/json/)
     expect(res.body).toHaveLength(1)
@@ -72,7 +107,7 @@ describe('sessions', () => {
     })
   })
   test('GET /:id', async () => {
-    const res = await api.get(`/api/sessions/${sess_id}`)
+    const res = await auth(api.get(`/api/sessions/${sess_id}`))
       .expect(200)
       .expect('Content-type', /application\/json/)
     expect(_.omit(res.body,['id']))
@@ -81,7 +116,7 @@ describe('sessions', () => {
   test('PUT /:id', async () => {
     userSession.name = 'new name'
     userSession.imgs.push({...userSession.imgs[0], scale: 1.2, left: 5})
-    const res = await api.put(`/api/sessions/${sess_id}`)
+    const res = await auth(api.put(`/api/sessions/${sess_id}`))
       .send(userSession)
       .expect(200)
       .expect('Content-type', /application\/json/)
@@ -89,14 +124,14 @@ describe('sessions', () => {
       .toEqual(userSession);
   })
   test('DELETE /:id', async () => {
-    await api.delete(`/api/sessions/${sess_id}`)
+    await auth(api.delete(`/api/sessions/${sess_id}`))
       .expect(204);
   })
   test('inexistent sessions', async () => {
     const url = '/api/sessions/123456789012345678901234'
-    await api.get(url).expect(404)
-    await api.delete(url).expect(404)
-    await api.put(url).send(userSession).expect(404)
+    await auth(api.get(url)).expect(404)
+    await auth(api.delete(url)).expect(404)
+    await auth(api.put(url).send(userSession)).expect(404)
   })
 })
